+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision History
 
+## v ?.?.? - ?? ??? ????
+- TextureUtils::loadAndRegister2DTexture() can silence error message
+- MD5Model silences all but final texture fail messages during load
+
 ## v 5.4.1 - 26 Aug 2024
 - Updated README